Minor re-organization of new classes and interfaces.
